requirements.txt
# Appium自动化测试框架依赖包

# 核心Appium相关
Appium-Python-Client>=3.1.0
selenium>=4.15.0

# 测试框架
pytest>=7.4.0
pytest-html>=4.0.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0.0

# 报告和日志
allure-pytest>=2.13.0
colorlog>=6.7.0

# 数据处理
PyYAML>=6.0
pandas>=2.1.0
openpyxl>=3.1.0

# 图像处理（用于图像识别和截图对比）
Pillow>=10.0.0
opencv-python>=4.8.0

# HTTP请求（用于接口测试）
requests>=2.31.0

# 工具库
retry>=0.9.2
jsonschema>=4.19.0

# 开发工具
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0